chore: move and rename files to improve organization.
