+++ packages/shared/common/src/index.ts
@@ -7,6 +7,7 @@ export * from './validators';
 export * from './logging';
 export * from './options';
 export * from './utils';
+export * from './Context';
 
 export * as internal from './internal';
 export * from './errors';
